Move example composites into tests.
