+++ myia/operations/primitives.py
@@ -12,12 +12,7 @@ subclass.
 # The script will search for all primitives it can find in myia.operations    #
 ###############################################################################
 
-from .utils import (
-    BackendPrimitive,
-    CompositePrimitive,
-    InferencePrimitive,
-    PlaceholderPrimitive,
-)
+from .utils import BackendPrimitive, InferencePrimitive, PlaceholderPrimitive
 
 J = PlaceholderPrimitive(name="J", defaults="myia.operations.prim_J")
 
@@ -79,14 +74,6 @@ broadcast_shape = BackendPrimitive(
 
 casttag = BackendPrimitive(
     name="casttag", defaults="myia.operations.prim_casttag"
-)
-
-composite_full = CompositePrimitive(
-    name="composite_full", defaults="myia.operations.prim_composite_full"
-)
-
-composite_simple = CompositePrimitive(
-    name="composite_simple", defaults="myia.operations.prim_composite_simple"
 )
 
 concat = BackendPrimitive(name="concat", defaults="myia.operations.prim_concat")
